Choose the cron runner from env metadata
By default, the `cron` cookbook installs the customer-defined
jobs on the app master (or solo, as the case may be).

That's fine default behavior, but most want to move those jobs
off of app_master and onto a util.

With this change, the jobs defined for an environment will be
installed on the instance named in the `crontab_instance_name`
environment metadata.

Example:

`{"crontab_instance_name":"Zappity"}` in the environment metadata
will install the dashboard-defined cron jobs only on the utility
instance named "Zappity"