db-ssl timeouts and corrections
The `remote_key_copy.sh` script that is generated, when there are
stopped instances in the environment, causes chef runs to take
upwards of nine extra minutes per stopped instance. This is because
the `push_keys` function in this script, which uses `rsync` to copy
a few key files to each instance in the environment, and the
connection timeout is around 3 minutes by default.

Additionally, the `push_keys` function is used (contextually) to
copy those same files to the db replicas in the environment, but
this is not actually happening due to the use of global variables.

Changes:

* Defined an `is_up` function that takes an instance as an argument.
  It then attempts to ssh to the instance and run `date`. If this
  should fail, it returns false. Otherwise, it is true. Either way
  it has a hard-coded ten-second connection timeout.
* Modified `push_keys` to take an instance argument, treating it
  as the target for the `rsync` calls.
* Modified all `push_keys` and raw `rsync` invocations to first
  check to ensure that the instance in question `is_up`.